move TableMetaData and ColumnMetaData to shardingsphere-sql-parser-binder module
