+++ sharding-core/sharding-core-entry/src/test/java/org/apache/shardingsphere/core/shard/fixture/RoutingHookFixture.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.core.shard.fixture;
 
-import org.apache.shardingsphere.underlying.common.metadata.table.TableMetas;
+import org.apache.shardingsphere.sql.parser.binder.metadata.table.TableMetas;
 import org.apache.shardingsphere.sharding.route.engine.context.ShardingRouteContext;
 import org.apache.shardingsphere.sharding.route.hook.RoutingHook;
