+++ sharding-core/sharding-core-route/src/main/java/org/apache/shardingsphere/sharding/route/hook/SPIRoutingHook.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.sharding.route.hook;
 
-import org.apache.shardingsphere.underlying.common.metadata.table.TableMetas;
+import org.apache.shardingsphere.sql.parser.binder.metadata.table.TableMetas;
 import org.apache.shardingsphere.sharding.route.engine.context.ShardingRouteContext;
 import org.apache.shardingsphere.spi.NewInstanceServiceLoader;
