+++ shardingsphere-sql-parser/shardingsphere-sql-parser-binder/src/main/java/org/apache/shardingsphere/sql/parser/binder/metadata/column/ColumnMetaData.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.underlying.common.metadata.column;
+package org.apache.shardingsphere.sql.parser.binder.metadata.column;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
